parser/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0

# PDF processing
PyPDF2==3.0.1

# LLM and AI
phidata==2.4.25
google-generativeai>=0.7.0

# Data validation
pydantic==2.5.0

# Development
python-dotenv==1.0.0
